fix: cast retry backoff operands for CockroachDB

+++ src/plans.ts
@@ -1090,8 +1090,8 @@ function failJobs (schema: string, table: string, where: string, output: string)
         CASE WHEN retry_count = retry_limit THEN start_after
              WHEN NOT retry_backoff THEN now() + retry_delay * interval '1'
              ELSE now() + LEAST(
-               retry_delay_max,
-               retry_delay * (
+               retry_delay_max::float8,
+               retry_delay::float8 * (
                 2 ^ LEAST(16, retry_count + 1) / 2 +
                 2 ^ LEAST(16, retry_count + 1) / 2 * random()
                )
